// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/ClubRepRegistrationRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClubRepRegistrationRequest = {
    clubName: string;
    firstName: string;
    lastName: string;
    email: string;
    username: string;
    password: string;
    streetAddress: string;
    city: string;
    state: string;
    postalCode: string;
    cellphone: string;
};